requirements.txt
betterproto==1.2.5
certifi==2023.7.22
charset-normalizer==3.2.0
grpclib==0.4.5
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
multidict==6.0.4
requests==2.31.0
stringcase==1.2.0
urllib3==2.0.4
websocket-client==1.6.1
